Move `inspectValue()` to `LogRecord`.
